writing tests and classes
I've started and nearly completed these classes:
CodePegs
DecodingBoard

I started KeyPegs and realized that the key peg set might be better
represented as a hash and the same with the code peg set. It should be a
trivial change but necessary to do soon.
I've decided definitely in favor of creating a GuessEvaluator class as
well as a GameFlow class although I haven't done it yet.