changing structure of peg classes
after some though I decided I wanted the peg classes to take on peg
creation and peg set creation as well as peg validation.

I made the necessary changes and tested them. I'm satisfied with the
results. the decodingboard is already up to snuff.

after the changes to the peg classes I don't really need a codemaker
class. I kept it in case I want to use that for player agency - so they
can create their own code and play against another person or even have
the computer try to guess their code. that might be a fun thing to do
sometime.

I've started on the codebreaker class. I've identified 4 methods I think
it should have. It seems that I will not need a GuessEvaluator class
afterall. CodeBreaker can handle it.

So, I'll write that. It seems like it will need to be aware of the peg
classes and the decodingboard class.  It will need to notify the
GameFlow (or whatever) class once all of the pages are identified.

GameFlow class will likely track guess attempts and track victory/loss.
This class will have our game loop.