CodeBreaker
this is a difficult class to work on. It'd be easy to just have it make
an all new guess everytime - and that would still fit the parameters of
the lesson - but it feels dishonest. Making the program do intelligent
guesses is difficult. Nevertheless, progress has been made.
There are definite issues with the heavy reliance on arrays. Ruby syntax
prefers hashes.
More work to be done here, and then solid tests.
CodeMaker should be easy after this.